Finish clean architecture refactor
