+++ app/src/main/java/com/maximo/douglas/githubconsumer/Application.kt
@@ -14,10 +14,11 @@ class Application : Application() {
         startKoin {
             modules(
                 listOf(
-                    viewModelModule,
-                    repositoryModule,
                     commonsModule,
-                    useCaseModule
+                    serviceModule,
+                    repositoryModule,
+                    useCaseModule,
+                    viewModelModule
                 )
             )
         }
