+++ app/src/test/java/com/maximo/douglas/githubconsumer/viewmodels/gitrepository/context/SelectGitRepositoryViewModelTest.kt
@@ -2,8 +2,8 @@ package com.maximo.douglas.githubconsumer.viewmodels.gitrepository.context
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
-import com.maximo.douglas.data.repository.GitRepositoryRepository
 import com.maximo.douglas.domain.entity.gitrepository.GitRepository
+import com.maximo.douglas.domain.usecase.GetGitRepositoryUseCase
 import com.maximo.douglas.githubconsumer.testutils.TestContextProvider
 import com.maximo.douglas.githubconsumer.testutils.TestCoroutineRule
 import com.maximo.douglas.githubconsumer.testutils.faker.GitRepositoryFaker
@@ -27,7 +27,7 @@ open class SelectGitRepositoryViewModelTest {
     val testCoroutineRule = TestCoroutineRule()
 
     @Mock
-    private lateinit var mGitRepositoryRepository: GitRepositoryRepository
+    private lateinit var mGetGitRepositoryUseCase: GetGitRepositoryUseCase
 
     @Mock
     private lateinit var mViewStateObserver: Observer<SelectGitRepositoryViewModelState>
@@ -47,7 +47,7 @@ open class SelectGitRepositoryViewModelTest {
         MockitoAnnotations.initMocks(this)
 
         mSelectRepositoryViewModel = SelectGitRepositoryViewModel(
-            gitRepositoryRemoteDataSource = mGitRepositoryRepository,
+            getGitRepositoryUseCase = mGetGitRepositoryUseCase,
             contextProvider = TestContextProvider()
         ).apply {
             getStateLiveData().observeForever(mViewStateObserver)
@@ -64,7 +64,7 @@ open class SelectGitRepositoryViewModelTest {
 
     fun `should success when data source returns the expected data`() =
         testCoroutineRule.runBlockingTest {
-            `when`(mGitRepositoryRepository.getGitRepositoryList(REQUEST_PAGE)).thenReturn(
+            `when`(mGetGitRepositoryUseCase.invoke(REQUEST_PAGE)).thenReturn(
                 mGitRepositoryList
             )
 
@@ -81,7 +81,7 @@ open class SelectGitRepositoryViewModelTest {
     fun `should throw error when data source throws exception`() =
         testCoroutineRule.runBlockingTest {
             val error = Error()
-            `when`(mGitRepositoryRepository.getGitRepositoryList(REQUEST_PAGE)).thenThrow(
+            `when`(mGetGitRepositoryUseCase.invoke(REQUEST_PAGE)).thenThrow(
                 error
             )
