+++ data/src/main/java/com/maximo/douglas/data/di/ServiceModule.kt
@@ -1,12 +1,13 @@
 package com.maximo.douglas.data.di
 
-import com.maximo.douglas.data.remote.ApiDataSource.Companion.createService
-import com.maximo.douglas.data.remote.gitpullrequest.GitPullRequestApi
-import com.maximo.douglas.data.remote.gitrepository.GitRepositoryApi
+import com.maximo.douglas.commons.utils.ApiDataSource.Companion.createService
+import com.maximo.douglas.data.BuildConfig.BASE_URL
+import com.maximo.douglas.data.remote.GitPullRequestApi
+import com.maximo.douglas.data.remote.GitRepositoryApi
 import org.koin.dsl.module
 
 @Suppress("USELESS_CAST")
 val serviceModule = module {
-    single { createService(GitRepositoryApi::class.java) as GitRepositoryApi }
-    single { createService(GitPullRequestApi::class.java) as GitPullRequestApi }
+    single { createService(BASE_URL, GitRepositoryApi::class.java) as GitRepositoryApi }
+    single { createService(BASE_URL, GitPullRequestApi::class.java) as GitPullRequestApi }
 }
