--- data/src/main/java/com/maximo/douglas/data/remote/ApiDataSource.kt
@@ -1,39 +1,0 @@
-package com.maximo.douglas.data.remote
-
-import com.google.gson.GsonBuilder
-import com.maximo.douglas.data.BuildConfig.BASE_URL
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
-import retrofit2.converter.gson.GsonConverterFactory
-import java.util.concurrent.TimeUnit
-
-class ApiDataSource {
-
-    companion object {
-
-        fun <S> createService(serviceClass: Class<S>): S {
-            val loggingInterceptor = HttpLoggingInterceptor()
-            loggingInterceptor.level = HttpLoggingInterceptor.Level.BODY
-
-            val httpClient = OkHttpClient.Builder()
-                .readTimeout(15, TimeUnit.SECONDS)
-
-            httpClient.addInterceptor(loggingInterceptor)
-
-            val gson = GsonBuilder()
-                .setDateFormat("yyyy-MM-dd HH:mm:ss")
-                .create()
-
-            val retrofit = Retrofit.Builder()
-                .baseUrl(BASE_URL)
-                .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
-                .addConverterFactory(GsonConverterFactory.create(gson))
-                .client(httpClient.build())
-                .build()
-
-            return retrofit.create(serviceClass)
-        }
-    }
-}
