+++ domain/build.gradle
@@ -61,7 +61,6 @@ android {
 
 dependencies {
     implementation project(":commons")
-    implementation project(":data")
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'androidx.core:core-ktx:1.3.2'
@@ -75,4 +74,6 @@ dependencies {
     // KOIN
     implementation "io.insert-koin:koin-android:3.0.1"
     testImplementation "io.insert-koin:koin-test-junit4:3.0.1"
+
+    implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
 }
