+++ domain/src/main/java/com/maximo/douglas/domain/entity/gitpullrequest/GitPullRequest.kt
@@ -23,3 +23,8 @@ data class GitPullRequest(
     }
 
 }
+
+enum class PullRequestStateEnum(val value: String) {
+    OPEN("open"),
+    CLOSED("closed")
+}
