--- domain/src/main/java/com/maximo/douglas/domain/entity/gitpullrequest/PullRequestStateEnum.kt
@@ -1,6 +1,0 @@
-package com.maximo.douglas.domain.entity.gitpullrequest
-
-enum class PullRequestStateEnum(val value: String) {
-    OPEN("open"),
-    CLOSED("closed")
-}
