+++ domain/src/main/java/com/maximo/douglas/domain/repository/GitRepositoryRepository.kt
@@ -1,0 +1,11 @@
+package com.maximo.douglas.domain.repository
+
+import com.maximo.douglas.domain.entity.gitrepository.GitRepository
+
+interface GitRepositoryRepository {
+
+    suspend fun getGitRepositoryList(
+        page: Int
+    ): List<GitRepository>
+
+}
